Module4/requirements.txt
dash
plotly.express
pandas
dash_bootstrap_components
dash_bootstrap_templates